+++ src/main/java/io/jenkins/plugins/analysis/warnings/GnuFortran.java
@@ -39,11 +39,6 @@ public class GnuFortran extends ReportScanningTool {
             super(ID);
         }
 
-        @Override
-        public String getHelp() {
-            return Messages.Warning_SlowMultiLineParser();
-        }
-
         @NonNull
         @Override
         public String getDisplayName() {