+++ src/test/java/io/jenkins/plugins/analysis/warnings/StepsITest.java
@@ -10,12 +10,12 @@ import java.util.stream.Collectors;
 
 import org.eclipse.collections.impl.factory.Lists;
 import org.junit.Assume;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.jvnet.hudson.test.TestExtension;
 
 import edu.hm.hafner.analysis.Issue;
 import edu.hm.hafner.analysis.Report;
+import edu.hm.hafner.analysis.Severity;
 
 import org.kohsuke.stapler.HttpResponse;
 import org.jenkinsci.Symbol;
@@ -59,7 +59,7 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
     /**
      * Creates a declarative Pipeline and scans for a Gcc warning.
      */
-    @Test @Ignore("See https://github.com/jenkinsci/pipeline-model-definition-plugin/pull/314")
+    @Test
     public void shouldRunInDeclarativePipeline() {
         WorkflowJob job = createPipeline();
 
@@ -68,7 +68,7 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
                 + "    stages {\n"
                 + "        stage ('Create a fake warning') {\n"
                 + "            steps {\n"
-                + "                sh 'echo \"foo.cc:4:39: error: foo.h: No such file or directory\" >warnings.log' "
+                + getShellStep("echo \"foo.cc:4:39: error: foo.h: No such file or directory\" >warnings.log")
                 + "            }\n"
                 + "        }\n"
                 + "    }\n"
@@ -82,6 +82,15 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
         AnalysisResult result = scheduleSuccessfulBuild(job);
 
         assertThat(result).hasTotalSize(1);
+    }
+
+    private String getShellStep(final String script) {
+        if (isWindows()) {
+            return String.format("bat '%s'", script);
+        }
+        else {
+            return String.format("sh '%s'", script);
+        }
     }
 
     /**
@@ -114,6 +123,46 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
 
         assertThat(first.getResult().getIssues()).hasSize(5);
         assertThat(second.getResult().getIssues()).hasSize(3);
+    }
+
+    /** Runs the Clang parser on an output file that contains 1 issue. */
+    @Test
+    public void shouldFindAllClangIssuesIfConsoleIsAnnotatedWithTimeStamps() {
+        Assume.assumeFalse("Test not yet OS independent: requires UNIX commands", isWindows());
+
+        WorkflowJob job = createPipelineWithWorkspaceFiles("issue56484.txt");
+        job.setDefinition(asStage(
+                "sh 'cat *.txt'",
+                "def issues = scanForIssues tool: clang()",
+                PUBLISH_ISSUES_STEP));
+
+        AnalysisResult result = scheduleSuccessfulBuild(job);
+
+        assertThat(result).hasTotalSize(1);
+        assertThat(result.getIssues().get(0))
+                .hasLineStart(1)
+                .hasLineEnd(1)
+                .hasColumnStart(2)
+                .hasColumnEnd(2)
+                .hasMessage("This is an error.")
+                .hasFileName("test.c")
+                .hasSeverity(Severity.WARNING_HIGH);
+    }
+
+    /** Runs the Clang parser on an output file that contains 1 issue. */
+    @Test
+    public void shouldFindAllJavaIssuesIfConsoleIsAnnotatedWithTimeStamps() {
+        Assume.assumeFalse("Test not yet OS independent: requires UNIX commands", isWindows());
+
+        WorkflowJob job = createPipelineWithWorkspaceFiles("issue56484-maven.txt");
+        job.setDefinition(asStage(
+                "sh 'cat *.txt'",
+                "def issues = scanForIssues tool: java()",
+                PUBLISH_ISSUES_STEP));
+
+        AnalysisResult result = scheduleSuccessfulBuild(job);
+
+        assertThat(result).hasTotalSize(10);
     }
 
     /**
@@ -305,7 +354,7 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
     public void shouldShowWarningsOfGroovyParser() {
         WorkflowJob job = createPipelineWithWorkspaceFiles("pep8Test.txt");
         job.setDefinition(asStage(
-                "def groovy = scanForIssues " 
+                "def groovy = scanForIssues "
                         + "tool: groovyScript(parserId: 'groovy-pep8', pattern:'**/*issues.txt', reportEncoding:'UTF-8')",
                 "publishIssues issues:[groovy]"));
 
@@ -325,11 +374,11 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
         assertThat(result.getIssues()).hasSize(8);
         assertThat(result.getIssues().getPropertyCount(Issue::getOrigin)).containsOnly(entry(id, 8));
     }
-    
+
     /**
      * Registers a new {@link GroovyParser} (a Pep8 parser) in Jenkins global configuration and uses this parser twice.
      */
-    @Test 
+    @Test
     public void shouldUseGroovyParserTwice() {
         List<AnalysisResult> results = getAnalysisResults(runWith2GroovyParsers(false));
         assertThat(results).hasSize(2);
@@ -339,9 +388,9 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
     }
 
     /**
-     * Verifies that a warning will be logged if the user specified name and id <b>and not</b> {@code isAggregating}. 
+     * Verifies that a warning will be logged if the user specified name and id <b>and not</b> {@code isAggregating}.
      */
-    @Test 
+    @Test
     public void shouldLogWarningIfNameIsSetWhenNotAggregating() {
         List<AnalysisResult> results = getAnalysisResults(runWith2GroovyParsers(false,
                 "name: 'name'", "id: 'id'"));
@@ -359,11 +408,11 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
      * Registers a new {@link GroovyParser} (a Pep8 parser) in Jenkins global configuration and uses this parser twice.
      * Publishes the results into a single result.
      */
-    @Test 
+    @Test
     public void shouldUseGroovyParserTwiceAndAggregateIntoSingleResult() {
         List<AnalysisResult> results = getAnalysisResults(runWith2GroovyParsers(true));
         assertThat(results).hasSize(1);
-        
+
         AnalysisResult result = results.get(0);
         assertThat(result.getId()).isEqualTo("analysis");
     }
@@ -376,7 +425,7 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
     public void shouldUseGroovyParserTwiceAndAggregateIntoSingleResultWithCustomizableIdAndName() {
         Run<?, ?> build = runWith2GroovyParsers(true, "name: 'Custom Name'", "id: 'custom-id'");
         ResultAction action = getResultAction(build);
-        
+
         assertThat(action.getId()).isEqualTo("custom-id");
         assertThat(action.getDisplayName()).isEqualTo("Custom Name Warnings");
     }
@@ -389,7 +438,7 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
     public void shouldUseGroovyParserTwiceAndAggregateIntoSingleResultWithCustomizableName() {
         Run<?, ?> build = runWith2GroovyParsers(true, "name: 'Custom Name'");
         ResultAction action = getResultAction(build);
-        
+
         assertThat(action.getId()).isEqualTo("analysis");
         assertThat(action.getDisplayName()).isEqualTo("Custom Name Warnings");
     }
@@ -397,7 +446,7 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
     private Run<?, ?> runWith2GroovyParsers(final boolean isAggregating, final String... arguments) {
         WorkflowJob job = createPipelineWithWorkspaceFiles("pep8Test.txt");
         job.setDefinition(asStage(
-                "recordIssues aggregatingResults: " + isAggregating + ", tools: [" 
+                "recordIssues aggregatingResults: " + isAggregating + ", tools: ["
                         + "groovyScript(parserId:'groovy-pep8', pattern: '**/*issues.txt', id: 'groovy-1'),"
                         + "groovyScript(parserId:'groovy-pep8', pattern: '**/*issues.txt', id: 'groovy-2')"
                         + "] " + join(arguments)));
@@ -420,23 +469,18 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
         WorkflowJob job = createPipelineWithWorkspaceFiles("pmd-filtering.xml");
 
         setFilter(job, "includeFile('File.*.java')");
-        final Pmd tool = new Pmd();
         assertThat(scheduleSuccessfulBuild(job).getTotalSize()).isEqualTo(16);
 
         setFilter(job, "excludeFile('File.*.java')");
-        final Pmd tool1 = new Pmd();
         assertThat(scheduleSuccessfulBuild(job).getTotalSize()).isZero();
 
         setFilter(job, "includeFile('')");
-        final Pmd tool2 = new Pmd();
         assertThat(scheduleSuccessfulBuild(job).getTotalSize()).isEqualTo(16);
 
         setFilter(job, "excludeFile('')");
-        final Pmd tool3 = new Pmd();
         assertThat(scheduleSuccessfulBuild(job).getTotalSize()).isEqualTo(16);
 
         setFilter(job, "");
-        final Pmd tool4 = new Pmd();
         assertThat(scheduleSuccessfulBuild(job).getTotalSize()).isEqualTo(16);
 
         verifyIncludeFile(job, "File1.java");
@@ -456,7 +500,8 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
         AnalysisResult result = scheduleSuccessfulBuild(job);
         assertThat(result.getTotalSize()).isEqualTo(8 + 4);
 
-        setFilter(job, "includeFile('File1.java'), excludeCategory('Category1'), excludeType('Type1'), excludeNamespace('.*package1') ");
+        setFilter(job,
+                "includeFile('File1.java'), excludeCategory('Category1'), excludeType('Type1'), excludeNamespace('.*package1') ");
         AnalysisResult oneIssue = scheduleSuccessfulBuild(job);
         assertThat(oneIssue.getIssues().getFiles()).containsExactly("File1.java");
         assertThat(oneIssue.getIssues().getCategories()).containsExactly("Category2");
@@ -468,16 +513,17 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
         setFilter(job, "includeFile('" + fileName + "')");
 
         AnalysisResult result = scheduleSuccessfulBuild(job);
-        
+
         assertThat(result.getTotalSize()).isEqualTo(8);
         assertThat(result.getIssues().getFiles()).containsExactly(fileName);
     }
 
-    private void verifyExcludeFile(final WorkflowJob job, final String excludedFileName, final String expectedFileName) {
+    private void verifyExcludeFile(final WorkflowJob job, final String excludedFileName,
+            final String expectedFileName) {
         setFilter(job, "excludeFile('" + excludedFileName + "')");
 
         AnalysisResult result = scheduleSuccessfulBuild(job);
-        
+
         assertThat(result.getTotalSize()).isEqualTo(8);
         assertThat(result.getIssues().getFiles()).containsExactly(expectedFileName);
     }
@@ -550,7 +596,8 @@ public class StepsITest extends IntegrationTestWithJenkinsPerTest {
                     .get(YouCannotTriggerMe.class);
             assertThat(urlHandler).isNotNull();
 
-            assertThat(urlHandler.triggerCount).as("XXE detected for parser %s: URL has been triggered!", tool)
+            assertThat(urlHandler.triggerCount)
+                    .as("XXE detected for parser %s: URL has been triggered!", tool)
                     .isEqualTo(0);
         }
     }