+++ CHANGELOG.md
@@ -17,12 +17,22 @@ Additional new features:
  Support for [IBLinter](https://github.com/IBDecodable/IBLinter) - Xcode Storyboard Linter
 - [PR#21](https://github.com/jenkinsci/warnings-ng-plugin/pull/21):
  Support for [PHPStan](https://github.com/phpstan/phpstan) - PHP Static Analysis Tool 
+- [PR#132](https://github.com/jenkinsci/analysis-model/pull/132): 
+Added a parser for CMake warnings.
+- [PR#137](https://github.com/jenkinsci/analysis-model/pull/137):
+Added a parser for JSON output from Cargo.
 
 ### Fixed
 
 - Added missing dependency to JSON-Smart library.
+- [JENKINS-56484](https://issues.jenkins-ci.org/browse/JENKINS-56484),
+[analysis-model-PR#19](https://github.com/jenkinsci/analysis-model/pull/130): 
+Filter TimeStamper Console log decorations
 - [PR#19](https://github.com/jenkinsci/warnings-ng-plugin/pull/19): Use `BuildStepMonitor.NONE` in step 
 to allow for higher concurrency.
+- [JENKINS-56324](https://issues.jenkins-ci.org/browse/JENKINS-56324): 
+If an issue in the current build is equal to several issues in the reference build 
+then select the issue with the same fingerprint as the outstanding issue. 
 
 ### Changed